requirements.txt
langchain
langchain-google-genai
langchain-chroma
langgraph
gradio
chromadb
pypdf
python-dotenv
langchain-community
langchain-text-splitters
fastapi
uvicorn
